+++ utils/tests/utils.go
@@ -3,11 +3,6 @@ package tests
 import (
 	"errors"
 	"fmt"
-	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
-	"github.com/jfrog/jfrog-client-go/artifactory"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"strings"
@@ -147,18 +142,4 @@ func CompareTree(a, b *services.GraphNode) bool {
 		}
 	}
 	return true
-}
-
-type restsTestHandler func(w http.ResponseWriter, r *http.Request)
-
-// Create mock server to test REST APIs.
-// t           - The testing object
-// testHandler - The HTTP handler of the test
-func CreateRestsMockServer(t *testing.T, testHandler restsTestHandler) (*httptest.Server, *config.ServerDetails, artifactory.ArtifactoryServicesManager) {
-	testServer := httptest.NewServer(http.HandlerFunc(testHandler))
-	serverDetails := &config.ServerDetails{ArtifactoryUrl: testServer.URL + "/"}
-
-	serviceManager, err := utils.CreateServiceManager(serverDetails, -1, 0, false)
-	assert.NoError(t, err)
-	return testServer, serverDetails, serviceManager
 }
